refactor(Background): move off background container from Board/index.js file
